Bug fixes
Fixed Rounding problem where truncation was happening before being cast for Math.Ceiling.

Reworking ColorToString as well as added debug.log entries to track data as it moves through.
Final decrypt algo still incorrect but closer to resolved.

+++ ColourCodes/Assets/Code/Encrypt.cs
@@ -324,7 +324,7 @@ namespace Evernorth.ColourCodes
         {
             string tempString = "";
             int sPos = 0;
-            int sLength = (int)Math.Ceiling((double)(s.Length / 2));
+            int sLength = (int)Math.Ceiling((decimal)(s.Length / 2));
 
             //Debug.Log(
             //    $"s.Length: {s.Length}\n" +
